Instead of storing the visible sandwich indices as state, compute it dynamically on demand. Add unstar/star buttons to the DraggableSandwich component.
This allows us to have render() perform the final computation for which
sandwiches are visible, and only store the minimal info as component
state.

That means when the starred state (stored in a React Context) is
updated, the rendered component will always reflect the latest starred
state (since we recalculate in render() instead of on button clicks).